refactor: move donations routes to new folder structure
