+++ src/API/REST/V3/Routes/Donations/RegisterDonationRoutes.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Give\Donations\Routes;
+namespace Give\API\REST\V3\Routes\Donations;
 
+use Give\API\REST\V3\Routes\Donations\ValueObjects\DonationRoute;
 use Give\Donations\Controllers\DonationRequestController;
 use Give\Donations\ValueObjects\DonationAnonymousMode;
-use Give\Donations\ValueObjects\DonationRoute;
 use WP_Error;
 use WP_REST_Request;
 use WP_REST_Server;
