+++ src/Donations/Controllers/DonationRequestController.php
@@ -2,9 +2,9 @@
 
 namespace Give\Donations\Controllers;
 
+use Give\API\REST\V3\Routes\Donations\ValueObjects\DonationRoute;
 use Give\Donations\Models\Donation;
 use Give\Donations\ValueObjects\DonationAnonymousMode;
-use Give\Donations\ValueObjects\DonationRoute;
 use WP_Error;
 use WP_REST_Request;
 use WP_REST_Response;
