+++ tests/Unit/API/REST/V3/Routes/Donations/GetDonationRouteTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Unit\Donations\Routes;
+namespace Unit\API\REST\V3\Routes\Donations;
 
 use Exception;
+use Give\API\REST\V3\Routes\Donations\RegisterDonationRoutes;
+use Give\API\REST\V3\Routes\Donations\ValueObjects\DonationRoute;
 use Give\Donations\Models\Donation;
-use Give\Donations\Routes\RegisterDonationRoutes;
-use Give\Donations\ValueObjects\DonationRoute;
 use Give\Donations\ValueObjects\DonationStatus;
 use Give\Helpers\Hooks;
 use Give\Tests\RestApiTestCase;
